+++ hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/map/MapSetWithMaxIdleMessageTask.java
@@ -19,6 +19,7 @@ package com.hazelcast.client.impl.protocol.task.map;
 import com.hazelcast.client.impl.protocol.ClientMessage;
 import com.hazelcast.client.impl.protocol.codec.MapSetWithMaxIdleCodec;
 import com.hazelcast.instance.Node;
+import com.hazelcast.internal.cluster.Versions;
 import com.hazelcast.map.impl.operation.MapOperation;
 import com.hazelcast.map.impl.operation.MapOperationProvider;
 import com.hazelcast.nio.Connection;
@@ -27,10 +28,16 @@ import com.hazelcast.spi.Operation;
 import java.util.concurrent.TimeUnit;
 
 public class MapSetWithMaxIdleMessageTask
-        extends AbstractMapSetWithMaxIdleMessageTask<MapSetWithMaxIdleCodec.RequestParameters> {
+        extends AbstractMapSetMessageTask<MapSetWithMaxIdleCodec.RequestParameters> {
 
     public MapSetWithMaxIdleMessageTask(ClientMessage clientMessage, Node node, Connection connection) {
         super(clientMessage, node, connection);
+    }
+
+    @Override
+    protected void beforeProcess() {
+        super.beforeProcess();
+        checkCompatibility();
     }
 
     @Override
@@ -65,5 +72,11 @@ public class MapSetWithMaxIdleMessageTask
     public Object[] getParameters() {
         return new Object[]{parameters.key, parameters.value, parameters.ttl, TimeUnit.MILLISECONDS,
                 parameters.maxIdle, TimeUnit.MILLISECONDS};
+    }
+
+    private void checkCompatibility() {
+        if (nodeEngine.getClusterService().getClusterVersion().isLessThan(Versions.V3_11)) {
+            throw new UnsupportedOperationException("Setting MaxIdle is available when cluster version is 3.11 or higher");
+        }
     }
 }
